网络消息处理

+++ apps/lib-core/src/main/kotlin/com/lj/core/net/msg/Handler.kt
@@ -1,4 +1,4 @@
 package com.lj.core.net.msg
 @Target(AnnotationTarget.CLASS)
 @Retention(AnnotationRetention.RUNTIME)
-annotation class Handler(val opcode:Int)
+annotation class Handler(val opcode:Short)
